style updates galore - all due to cursor and my wise direction
