+++ src/app/home.module.css
@@ -195,26 +195,113 @@ input:checked + .slider:before {
 .posts {
   display: flex;
   flex-direction: column;
-  gap: 20px;
-  margin-top: 20px;
+  gap: 24px;
+  margin-top: 24px;
 }
 
 .post {
   background: white;
-  padding: 15px;
+  padding: 24px;
   box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-  transition: transform 0.2s ease;
   border: 1px solid #eee;
+  transition: all 0.2s ease;
 }
 
 .dark .post {
   background: #222;
-  box-shadow: 0 1px 3px rgba(255, 255, 255, 0.05);
   border: 1px solid #333;
+  box-shadow: 0 1px 3px rgba(255, 255, 255, 0.05);
 }
 
 .post:hover {
-  transform: translateY(-3px);
+  transform: translateY(-2px);
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.15);
+}
+
+.dark .post:hover {
+  box-shadow: 0 2px 4px rgba(255, 255, 255, 0.08);
+}
+
+.question {
+  font-size: 1.2rem;
+  font-weight: 500;
+  color: #1a1a1a;
+  margin-bottom: 16px;
+  line-height: 1.4;
+}
+
+.dark .question {
+  color: #fff;
+}
+
+.answer {
+  color: #444;
+  line-height: 1.6;
+  margin-bottom: 20px;
+  font-size: 1rem;
+  padding-left: 16px;
+  border-left: 3px solid #007bff;
+}
+
+.dark .answer {
+  color: #bbb;
+  border-left-color: #f36100;
+}
+
+.postMeta {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  padding-top: 16px;
+  border-top: 1px solid #eee;
+  color: #666;
+  font-size: 0.85rem;
+}
+
+.dark .postMeta {
+  border-top-color: #333;
+  color: #888;
+}
+
+.metaItem {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+}
+
+.modelBadge {
+  background: #e8f0fe;
+  color: #1967d2;
+  padding: 2px 8px;
+  font-size: 0.85rem;
+  font-weight: 500;
+}
+
+.dark .modelBadge {
+  background: #1a1a1a;
+  color: #3ea6ff;
+}
+
+.timestamp {
+  margin-left: auto;
+  color: #888;
+  font-size: 0.85rem;
+}
+
+.dark .timestamp {
+  color: #666;
+}
+
+.caller {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  color: #444;
+  font-weight: 500;
+}
+
+.dark .caller {
+  color: #bbb;
 }
 
 .postHeader {
@@ -235,15 +322,6 @@ input:checked + .slider:before {
 .postTitle {
   font-size: 1.2rem;
   font-weight: 600;
-}
-
-.postMeta {
-  font-size: 0.9rem;
-  color: #666;
-}
-
-.dark .postMeta {
-  color: #bbb;
 }
 
 .postContent {
@@ -277,15 +355,43 @@ input:checked + .slider:before {
 }
 
 .footer {
-  padding: 20px;
-  background: white;
-  color: #666;
+  padding: 40px 20px;
+  margin-top: 60px;
+  width: 100%;
+  background: transparent;
+  border-top: 1px solid #eee;
+  color: #888;
   text-align: center;
+  font-size: 0.85rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
 }
 
 .dark .footer {
-  background: #222;
+  border-top: 1px solid #333;
+  color: #666;
+}
+
+.footerLogo {
+  display: inline-flex;
+  align-items: center;
+  gap: 4px;
+  color: #1a1a1a;
+  font-weight: 500;
+}
+
+.dark .footerLogo {
   color: #bbb;
+}
+
+.footerAccent {
+  color: #f36100;
+}
+
+.footerAccent2 {
+  color: #007bff;
 }
 
 .createPostBtn {
@@ -385,8 +491,6 @@ input:checked + .slider:before {
   color: #3ea6ff;
 }
 
-
-
 @media (max-width: 1150px) {
   .navLinks {
     display: none;
@@ -415,7 +519,6 @@ input:checked + .slider:before {
     font-size: 0.9rem;
   }
 }
-
 
 @media (max-width: 680px) {
   .header {
