+++ src/components/Header.module.css
@@ -1,0 +1,305 @@
+.header {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background: white;
+  padding: 15px 40px;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+  position: fixed;
+  top: 0;
+  z-index: 100;
+  transition: background 0.3s ease, color 0.3s ease;
+}
+
+.header.dark {
+  background: #222;
+  box-shadow: 0 1px 3px rgba(255, 255, 255, 0.05);
+}
+
+.logo {
+  position: absolute;
+  left: 50%;
+  transform: translateX(-50%);
+  flex-grow: 1;
+  text-align: center;
+  font-size: 1.8rem;
+  font-weight: 600;
+  color: #000;
+  text-decoration: none;
+  transition: color 0.3s ease;
+}
+
+.header.dark .logo {
+  color: #eee;
+}
+
+.accent {
+  color: #f36100 !important;
+}
+
+.accent2 {
+  color: #007bff !important;
+}
+
+.navright {
+  display: flex;
+  gap: 20px;
+  padding: 5px 5px;
+  color: #555;
+  text-decoration: none;
+  font-size: 1rem;
+  transition: 0.3s;
+}
+
+.header.dark .navright {
+  color: #bbb;
+}
+
+.essentialControls {
+  display: flex;
+  align-items: center;
+  gap: 20px;
+}
+
+.toggleWrapper {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+}
+
+.switch {
+  position: relative;
+  display: inline-block;
+  width: 40px;
+  height: 20px;
+}
+
+.switch input {
+  opacity: 0;
+  width: 0;
+  height: 0;
+}
+
+.slider {
+  position: absolute;
+  cursor: pointer;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: #007bff;
+  transition: 0.4s;
+  border-radius: 20px;
+}
+
+.slider:before {
+  position: absolute;
+  content: "";
+  height: 16px;
+  width: 16px;
+  left: 2px;
+  bottom: 2px;
+  background-color: #ffffff;
+  transition: 0.4s;
+  border-radius: 50%;
+}
+
+input:checked + .slider {
+  background-color: #f36100;
+}
+
+input:checked + .slider:before {
+  transform: translateX(20px);
+}
+
+.createPostBtn {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+  padding: 8px 16px;
+  background: #007bff;
+  color: #fff;
+  border: none;
+  font-size: 1rem;
+  cursor: pointer;
+  transition: background 0.3s ease;
+}
+
+.dark .createPostBtn {
+  background: #f36100;
+}
+
+.createPostBtn:hover {
+  background: #f36100;
+}
+
+.dark .createPostBtn:hover {
+  background: #3ea6ff;
+}
+
+.navLinks {
+  display: flex;
+  align-items: center;
+  gap: 20px;
+}
+
+.signupBtn {
+  background: none;
+  border: none;
+  color: #555;
+  font-size: 1rem;
+  cursor: pointer;
+  transition: color 0.3s ease;
+}
+
+.header.dark .signupBtn {
+  color: #bbb;
+}
+
+.signupBtn:hover {
+  color: #007bff;
+}
+
+.dark .signupBtn:hover {
+  color: #3ea6ff;
+}
+
+.mobileMenuBtn {
+  display: none;
+  background: none;
+  border: none;
+  font-size: 1.5rem;
+  cursor: pointer;
+  color: #555;
+  padding: 5px;
+  transition: color 0.3s ease;
+}
+
+.header.dark .mobileMenuBtn {
+  color: #bbb;
+}
+
+.dropdownMenu {
+  position: absolute;
+  top: 100%;
+  right: 20px;
+  background: white;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+  padding: 10px;
+  min-width: 200px;
+  z-index: 101;
+  border: 1px solid #eee;
+  transition: background 0.3s ease, border-color 0.3s ease;
+}
+
+.header.dark .dropdownMenu {
+  background: #222;
+  box-shadow: 0 1px 3px rgba(255, 255, 255, 0.05);
+  border: 1px solid #333;
+}
+
+.dropdownItem {
+  display: block;
+  width: 100%;
+  padding: 10px 16px;
+  color: #555;
+  text-decoration: none;
+  font-size: 1rem;
+  transition: 0.3s;
+  border: none;
+  background: none;
+  text-align: left;
+  cursor: pointer;
+}
+
+.header.dark .dropdownItem {
+  color: #bbb;
+}
+
+.dropdownItem:hover {
+  background: rgba(0, 123, 255, 0.1);
+  color: #007bff;
+}
+
+.dark .dropdownItem:hover {
+  background: rgba(62, 166, 255, 0.1);
+  color: #3ea6ff;
+}
+
+@media (max-width: 1150px) {
+  .navLinks {
+    display: none;
+  }
+
+  .mobileMenuBtn {
+    display: block;
+  }
+
+  .dropdownMenu {
+    display: block;
+  }
+
+  .header {
+    padding: 15px 20px;
+  }
+}
+
+@media (max-width: 680px) {
+  .header {
+    position: relative;
+    padding: 15px 10px;
+  }
+
+  .essentialControls {
+    display: flex;
+    gap: 25px;
+    align-items: center;
+  }
+
+  .createPostBtn {
+    padding: 6px 10px;
+  }
+
+  .createPostText {
+    display: none;
+  }
+
+  .logo {
+    font-size: 1.6rem;
+    z-index: 1;
+  }
+
+  .switch {
+    transform: scale(0.9);
+  }
+}
+
+@media (max-width: 480px) {
+  .header {
+    padding: 15px 5px;
+  }
+
+  .logo {
+    font-size: 1.4rem;
+  }
+
+  .createPostBtn {
+    padding: 5px 8px;
+  }
+
+  .switch {
+    transform: scale(0.8);
+  }
+
+  .essentialControls {
+    gap: 5px;
+  }
+}
+
+@media (max-width: 390px) {
+  .logo {
+    position: relative;
+    left: 35%;
+  }
+}
